--- lua/_key_bindings.lua
@@ -1,1 +1,0 @@
-vim.g.mapleader = " "
